requirements.txt
snowflake-connector-python
python-dotenv
yfinance
faker
numpy
holidays